move some helper functions to helper.go
